+++ Appboy-tvOS-SDK/AppboyTVOSKit.framework/Headers/ABKBannerCard.h
@@ -1,7 +1,7 @@
 #import "ABKCard.h"
 
 /*
- * Appboy Public API: ABKBannerCard
+ * Braze Public API: ABKBannerCard
  */
 NS_ASSUME_NONNULL_BEGIN
 @interface ABKBannerCard : ABKCard <NSCoding>