+++ Appboy-tvOS-SDK/AppboyTVOSKit.framework/Headers/ABKFacebookUser.h
@@ -5,7 +5,7 @@ NS_ASSUME_NONNULL_BEGIN
 extern NSInteger const DefaultNumberOfFriends;
 
 /*
- * Appboy Public API: ABKFacebookUser
+ * Braze Public API: ABKFacebookUser
  */
 @interface ABKFacebookUser : NSObject
 
@@ -19,7 +19,7 @@ extern NSInteger const DefaultNumberOfFriends;
  * from facebook with facebook graph api endpoint "/me/likes", under the key "data"; Please refer to
  * https://developers.facebook.com/docs/graph-api/reference/v2.2/user/likes for more information.
  *
- * @discussion: This method is to generate a ABKFacebookUser so you can pass the user's facebook account data to Appboy.
+ * @discussion: This method is to generate a ABKFacebookUser so you can pass the user's facebook account data to Braze.
  * After a ABKFacebookUser object is generated, you can check the value of properties but you cannot change it.
  * If you want to update the user's facebook data, you need to generate a new ABKFacebookUser instance and set it as
  * [Appboy sharedInstance].user.facebookUser.