+++ Appboy-tvOS-SDK/AppboyTVOSKit.framework/Headers/ABKFeedController.h
@@ -50,13 +50,13 @@ typedef NS_OPTIONS(NSUInteger, ABKCardCategory) {
 };
 
 /*
- * Appboy Public API: ABKFeedController
+ * Braze Public API: ABKFeedController
  */
 @interface ABKFeedController : NSObject
 
 /*!
  * The latest cards of Appboy news feed that is saved in memory and disk. Right now the available card types are ABKBannerCard,
- * ABKCaptionedImageCard, ABKCrossPromotionCard, ABKClassicCard and ABKTextAnnouncementCard. They are all subclasses
+ * ABKCaptionedImageCard, ABKClassicCard and ABKTextAnnouncementCard. They are all subclasses
  * of ABKCard.
  */
 @property (readonly, getter=getNewsFeedCards) NSArray *newsFeedCards;