+++ AppboyKit/headers/AppboyKitLibrary/AppboyKit.h
@@ -1,6 +1,5 @@
 #import "Appboy.h"
 #import "ABKUser.h"
-#import "ABKAppboyEndpointDelegate.h"
 #import "ABKFacebookUser.h"
 #import "ABKTwitterUser.h"
 #import "ABKAttributionData.h"